Move main source file to src/
- updated gitignore
- fixed headers for tests
