+++ libpam/src/hmac.c
