+++ libpam/src/hmac.h
