+++ libpam/src/sha1.c
